Split up sitemap.rb
`sitemap.rb` currently lives in `lib/elasticsearch`, where is doesn't
really belong. This commit splits up `sitemap.rb` into a file per
class, moves the lib files to `lib/sitemap`, the tests to
`test/unit/sitemap`.
